Refactor datatypes
