+++ projector-html/src/Projector/Html/Backend.hs
@@ -1,29 +1,18 @@
 {-# LANGUAGE NoImplicitPrelude #-}
 {-# LANGUAGE OverloadedStrings #-}
 module Projector.Html.Backend (
-    BackendT (..)
+    Backend
+  , BackendT (..)
   , getBackend
-  , Backend (..)
   , haskellBackend
   , purescriptBackend
   ) where
 
 
-import           P
-
-import           Projector.Core  (Name)
-import           Projector.Html.Backend.Data
+import           Projector.Html.Data.Backend
 import qualified Projector.Html.Backend.Haskell as Hs
 import qualified Projector.Html.Backend.Purescript as Purs
-import           Projector.Html.Core
 
-import           System.IO  (FilePath)
-
-
-data BackendT
-  = Haskell
-  | Purescript
-  deriving (Eq, Ord, Show)
 
 getBackend :: BackendT -> Backend a
 getBackend b =
@@ -33,13 +22,6 @@ getBackend b =
 
     Purescript ->
       purescriptBackend
-
--- -----------------------------------------------------------------------------
-
-data Backend a = Backend {
-    renderModule :: ModuleName -> Module a -> (FilePath, Text)
-  , renderExpr :: Name -> HtmlExpr a -> Text
-  }
 
 haskellBackend :: Backend a
 haskellBackend =
